Web App for homeless dogs.
App users can send information about street dogs at any time through their app. The system includes three components namely the mobile app that is used by the community, the mobile app used by the authorized personnel at Dogee, and the web portal which has all data of rescued street dogs. All the data will be controlled by a centralized web server.

We need to develop a mobile app for making a rescue request, being a foster care for a street dog, making donations, being a sponsor for a street dog.

And also, we need to develop a web portal to use by the Dogee organization. It contains all data related to processes which are rescuing, hospitalization, adoption. It also contains a separate space to view the ongoing rescue requests.